Fixed: design of the pie charts
